Add: Add endpoint for grid overload and chat integration
This change adds an endpoint (/api/critical) that can be used by the
strapi CMS to trigger messages before potential grid overloads. It
responds to this endpoint by sending a message in the chat box and
asking the utility user to confirm whether he wishes to release a DFP
program or not. Currrently this is just a placeholder, later this can
probably trigger a request